Make control unit sequential again
This goes back to Verilog from SystemVerilog.

Why?

Verilator (rightly) complained that the latches in the control unit were
implicit. The easiest way to make them explicit latches would be to use
`always_latch`, which is a SystemVerilog construct. So I switched to
that.
Now you would think that the industry-standard FPGA toolchain, AMD
Vivado, would support SystemVerilog modules as the top module for a
design. Apparently not.
So now I had this design that would work in only one place. Sure, I
could write a Verilog wrapper for the SystemVerilog module that Vivado
would accept, but this project is janky enough as it is.
I didn't want to deal with more control unit race conditions, so I
decided to make it sequential, replacing all the implicit latches with
FFs. IPC is lower now, but the code is cleaner. I probably should've
done this all along.

Also, there were some random registers that I added forever ago that I
had long since stopped using but never actually deleted. Finally found
(some?) of those and deleted them.